Attempt to simplify the token cache providers DO NOT MERGE
Ideally we'd want the implementation of the token cache
providers to be simplified. However this is not yet possible
because the MSAL Abstract token cache provider contains a boolean telling
if the token cache is an app or user token cache (so that the key
can be computed).

Ideally we'd want the ITokenCache to tell the application if it's an app token cache or a user token cache.