Replaced roles claim type
Azure AD returns a token with the new claim type for an app registration issued for a permission of type "Application". It is the case when the API that uses `Microsoft.Identity.Web.WebApiServiceCollectionExtensions.AddProtectedWebApi` is called by a daemon app without user login (following [this documentation](https://docs.microsoft.com/en-us/azure/active-directory/develop/scenario-protected-web-api-app-registration#if-your-web-api-is-called-by-a-daemon-app)).